Improvements to tests
